fdl:src: fdl function and code format
  adjust code format
